+++ neo/translations/de.php
@@ -1,0 +1,16 @@
+<?php
+
+return [
+	"Add block above" => "Block oberhalb hinzuf\xC3\xBCgen",
+	"Block type" => "Blocktyp",
+	"The maximum number of blocks of this type the field is allowed to have." => "Die maximale Anzahl an Blocks des Types, die dieses Feld besitzen darf.",
+	"Child Blocks" => "Kind-Bl\xC3\xB6cke",
+	"Which block types do you want to allow as children?" => "Welche Blocktypen m\xC3\xB6chtest Du als Kinder erlauben?",
+	"Delete block type" => "Blocktyp l\xC3\xB6schen",
+	"Delete group" => "Gruppe l\xC3\xB6schen",
+	"Top Level" => "H\xC3\xB6chste Ebene",
+	"Will this block type be allowed at the top level?" => "Soll dieser Blocktyp f\xC3\xBCr die h\xC3\xB6chste Ebene erlaubt sein?",
+	"Define the types of blocks that can be created within this Neo field, as well as the fields each block type is made up of." => "Definiere die Blocktypen, die dieses Neo-Feld haben darf und die Felder, aus denen jeder Block besteht.",
+	"Unable to nest Neo fields." => "Neo-Felder k\xC3\xB6nnen nicht verschachtelt werden.",
+	"Duplicate block" => "Doppelblock",
+];